Smaller refactors based on reviews
